Task21 completed

+++ src/main/java/com/epam/jf/common/homework/Task21.java
@@ -1,5 +1,6 @@
 package com.epam.jf.common.homework;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -15,7 +16,7 @@ public abstract class Task21 {
      * @param measurements Измерения в ходе эксперимента.
      * @return Вычисленное по исходным данным сопротивление.
      */
-    public abstract double calcResistance(List<IMeasurement> measurements);
+    public abstract BigDecimal calcResistance(List<IMeasurement> measurements);
 
     /** Измерение. */
     public interface IMeasurement {
